extract functions to new file and adapt to requests
